+++ src/pages/OrderPage/components/ReceiverSection/NoReceiver/NoReceiver.styles.tsx
@@ -1,0 +1,21 @@
+import styled from "@emotion/styled";
+
+export const NoReceiversContent = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: ${({ theme }) => theme.spacing[1]};
+
+  height: 100%;
+  border: 1px solid ${({ theme }) => theme.colors.gray[300]};
+  border-radius: ${({ theme }) => theme.borderRadius.sm};
+  padding: ${({ theme }) => theme.spacing[8]};
+`;
+
+export const NoReceiversLabel = styled.p`
+  font-size: ${({ theme }) => theme.typography.label.label1Regular.fontSize};
+  font-weight: ${({ theme }) =>
+    theme.typography.label.label1Regular.fontWeight};
+  color: ${({ theme }) => theme.colors.gray[600]};
+`;
