+++ src/pages/OrderPage/components/ReceiverSection/ReceiverModal/ReceiverModal.styles.tsx
@@ -1,31 +1,5 @@
 import styled from "@emotion/styled";
 
-export const ReceiverSectionHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-export const NoReceiversLabel = styled.p`
-  font-size: ${({ theme }) => theme.typography.label.label1Regular.fontSize};
-  font-weight: ${({ theme }) =>
-    theme.typography.label.label1Regular.fontWeight};
-  color: ${({ theme }) => theme.colors.gray[600]};
-`;
-
-export const NoReceiversContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: ${({ theme }) => theme.spacing[1]};
-
-  height: 100%;
-  border: 1px solid ${({ theme }) => theme.colors.gray[300]};
-  border-radius: ${({ theme }) => theme.borderRadius.sm};
-  padding: ${({ theme }) => theme.spacing[8]};
-`;
-
 export const ModalOverlay = styled.div`
   position: fixed;
   top: 0;
@@ -163,8 +137,9 @@ export const ReceiverInputHeader = styled.div`
   margin-bottom: ${({ theme }) => theme.spacing[2]};
 `;
 
-export const Divider = styled.div`
+export const Divider = styled.div<{ isLast: boolean }>`
   height: 1px;
   background-color: ${({ theme }) => theme.colors.gray[300]};
   margin-top: ${({ theme }) => theme.spacing[4]};
+  opacity: ${({ isLast }) => (isLast ? 0 : 1)};
 `;
